note that Random is currently broken
Due to caching the random to get a consistent random on multiple access, AND caching the wrapper in
the hybrid function, we will end up not being random. This needs thinking about.

Wrapping isn't a cheap operation, we can't abandon caching, and a future release should result in deep caching.