begin support for django style converting
previously, we would cast at every stage. This is not how Django does it.

Instead, Django does some minimal pre-processing during generation of an SQL statement,
it then runs conversion on the outermost expression.

We have the latter stage in. It needs cleaning up.
We also should reconsider how we do resolved_expressions. Right now, we're resolving at every stage.
We need to instead resolve at the beginning, and let it bubble up.